+++ sunpy/coordinates/frames.py
@@ -63,16 +63,16 @@ class HeliographicStonyhurst(BaseCoordinateFrame):
     >>> sc = SkyCoord(1*u.deg, 1*u.deg, 2*u.km,
     ...               frame="heliographic_stonyhurst",
     ...               obstime="2010/01/01T00:00:45")
-    >>> sc
-    <SkyCoord (HeliographicStonyhurst): obstime=2010-01-01 00:00:45,
-    lon=1.0 deg, lat=1.0 deg, rad=2.0 km>
-    >>> sc.frame
-    <HeliographicStonyhurst Coordinate: obstime=2010-01-01 00:00:45,
-    lon=1.0 deg, lat=1.0 deg, rad=2.0 km>
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (HeliographicStonyhurst: obstime=2010-01-01 00:00:45): (lon, lat, radius) in (deg, deg, km)
+        ( 1.,  1.,  2.)>
+    >>> sc.frame # doctest: +FLOAT_CMP
+    <HeliographicStonyhurst Coordinate (obstime=2010-01-01 00:00:45): (lon, lat, radius) in (deg, deg, km)
+        ( 1.,  1.,  2.)>
     >>> sc = SkyCoord(HeliographicStonyhurst(-10*u.deg, 2*u.deg))
-    >>> sc
-    <SkyCoord (HeliographicStonyhurst): obstime=None, lon=-10.0 deg,
-    lat=2.0 deg, rad=695508.0 km>
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (HeliographicStonyhurst: obstime=None): (lon, lat, radius) in (deg, deg, km)
+        (-10.,  2.,  695508.)>
 
     Notes
     -----
@@ -113,7 +113,7 @@ class HeliographicStonyhurst(BaseCoordinateFrame):
             elif isinstance(self._data, UnitSphericalRepresentation):
                 distance = RSUN_METERS.to(u.km)
 
-            if distance:
+            if distance is not None:
                 self._data = self.default_representation(lat=self._data.lat,
                                                          lon=self._data.lon,
                                                          distance=distance)
@@ -153,15 +153,15 @@ class HeliographicCarrington(HeliographicStonyhurst):
     >>> sc = SkyCoord(1*u.deg, 2*u.deg, 3*u.km,
     ...               frame="heliographic_carrington",
     ...               obstime="2010/01/01T00:00:30")
-    >>> sc
-    <SkyCoord (HelioGraphicCarrington): obstime=2010-01-01 00:00:30,
-    lon=1.0 deg, lat=2.0 deg, rad=3.0 km>
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (HeliographicCarrington: obstime=2010-01-01 00:00:30): (lon, lat, radius) in (deg, deg, km)
+        ( 1.,  2.,  3.)>
+
     >>> sc = SkyCoord([1,2,3]*u.deg, [4,5,6]*u.deg, [5,6,7]*u.km,
-    obstime="2010/01/01T00:00:45", frame="heliographic_carrington")
-    >>> sc
-    <SkyCoord (HelioGraphicCarrington): obstime=2010-01-01 00:00:45,
-    (lon, lat, rad) in (deg, deg, km)
-        [(1.0, 4.0, 5.0), (2.0, 5.0, 6.0), (3.0, 6.0, 7.0)]>
+    ...               obstime="2010/01/01T00:00:45", frame="heliographic_carrington")
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (HeliographicCarrington: obstime=2010-01-01 00:00:45): (lon, lat, radius) in (deg, deg, km)
+        [( 1.,  4.,  5.), ( 2.,  5.,  6.), ( 3.,  6.,  7.)]>
     """
 
     name = "heliographic_carrington"
@@ -212,20 +212,24 @@ class Heliocentric(BaseCoordinateFrame):
 
     Examples
     --------
+
     >>> from astropy.coordinates import SkyCoord, CartesianRepresentation
     >>> import sunpy.coordinates
     >>> import astropy.units as u
+
     >>> sc = SkyCoord(CartesianRepresentation(10*u.km, 1*u.km, 2*u.km),
     ...               obstime="2011/01/05T00:00:50", frame="heliocentric")
-    >>> sc
-    <SkyCoord (HelioCentric): obstime=2011-01-05 00:00:50, D0=149597870.7 km,
-    x=10.0 km, y=1.0 km, z=2.0 km>
-    >>> sc = SkyCoord([1,2]*u.km, [3,4]*u.m, [5,6]*u.cm, frame="heliocentric",
-    obstime="2011/01/01T00:00:54")
-    >>> sc
-    <SkyCoord (HelioCentric): obstime=2011-01-01 00:00:54, D0=149597870.7 km,
-    (x, y, z) in (km, m, cm)
-        [(1.0, 3.0, 5.0), (2.0, 4.0, 6.0)]>
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (Heliocentric: obstime=2011-01-05 00:00:50, observer=<HeliographicStonyhurst Coordinate (obstime=2011-01-05 00:00:50): (lon, lat, radius) in (deg, deg, AU)
+        ( 0., -3.43939103,  0.98334411)>): (x, y, z) in km
+        ( 10.,  1.,  2.)>
+
+    >>> sc = SkyCoord([1,2]*u.km, [3,4]*u.m, [5,6]*u.cm, frame="heliocentric", obstime="2011/01/01T00:00:54")
+
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (Heliocentric: obstime=2011-01-01 00:00:54, observer=<HeliographicStonyhurst Coordinate (obstime=2011-01-01 00:00:54): (lon, lat, radius) in (deg, deg, AU)
+        ( 0., -2.97725356,  0.98335586)>): (x, y, z) in (km, m, cm)
+        [( 1.,  3.,  5.), ( 2.,  4.,  6.)]>
     """
 
     default_representation = CartesianRepresentation
@@ -274,14 +278,15 @@ class Helioprojective(BaseCoordinateFrame):
     >>> import astropy.units as u
     >>> sc = SkyCoord(0*u.deg, 0*u.deg, 5*u.km, obstime="2010/01/01T00:00:00",
     ...               frame="helioprojective")
-    >>> sc
-    <SkyCoord (HelioProjective): obstime=2010-01-01 00:00:00, D0=149597870.7 km
-    , Tx=0.0 arcsec, Ty=0.0 arcsec, distance=5.0 km>
-    >>> sc = SkyCoord(0*u.deg, 0*u.deg, obstime="2010/01/01T00:00:00",
-    frame="helioprojective")
-    >>> sc
-    <SkyCoord (HelioProjective): obstime=2010-01-01 00:00:00, D0=149597870.7 km
-    , Tx=0.0 arcsec, Ty=0.0 arcsec, distance=149597870.7 km>
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (Helioprojective: obstime=2010-01-01 00:00:00, rsun=695508.0 km, observer=<HeliographicStonyhurst Coordinate (obstime=2010-01-01 00:00:00): (lon, lat, radius) in (deg, deg, AU)
+        ( 0., -3.00724817,  0.98330294)>): (Tx, Ty, distance) in (arcsec, arcsec, km)
+        ( 0.,  0.,  5.)>
+    >>> sc = SkyCoord(0*u.deg, 0*u.deg, obstime="2010/01/01T00:00:00", frame="helioprojective")
+    >>> sc # doctest: +FLOAT_CMP
+    <SkyCoord (Helioprojective: obstime=2010-01-01 00:00:00, rsun=695508.0 km, observer=<HeliographicStonyhurst Coordinate (obstime=2010-01-01 00:00:00): (lon, lat, radius) in (deg, deg, AU)
+        ( 0., -3.00724817,  0.98330294)>): (Tx, Ty) in arcsec
+        ( 0.,  0.)>
     """
 
     default_representation = SphericalWrap180Representation